Change casing on keywords, add Write-SnapshotMessage
Change casing on certain keywords to fall into line with PowerShell standards. (Yes, it's case invariant but standardizing makes it a bit easier to read.)

Also, added a Write-SnapshotMessage function to make the Snapshot message writing a bit more robust and cleaner to call from other scripts.

Removed the first check for the $snapshot variable because it was recreating the purpose of the switch type parameter.